app/requirements.txt
Django==3.1.12
djongo==1.3.7
mongoengine==0.20.0
pylint-mongoengine==0.4.0
pymongo==3.11.0
psycopg2-binary==2.8.5
djangorestframework==3.12.4
django-filter==2.4.0
django-cors-headers==3.9.0
pillow==10.4.0
